+++ R/CFAxisLongitude.R
@@ -10,6 +10,7 @@
 #' @export
 CFAxisLongitude <- R6::R6Class("CFAxisLongitude",
   inherit = CFAxisNumeric,
+  cloneable = FALSE,
   public = list(
 
     #' @description Create a new instance of this class.