+++ NEWS.md
@@ -1,6 +1,8 @@
 # ncdfCF (development version)
 - Dependency on R bumped from version 3.5 to 4.0.
 - With the `as_CFArray()` function you can create a CFArray instance from an R object such as a vector, matrix or array using logical, integer, double or character mode. Axes are created from dimnames, using names when set (such as in `dimnames(arr) <- list(X = 40:43, Y = 50:54, Z = 60:65)`) and possibly with latitude, longitude and time axes generated.
+- Functions from the Ops and Math group of S3 generic functions are now supported. This means that expressions can be written directly on CFArray instances.
+- All CFAxis descendant classes now have a `copy()` method which creates a deep copy of the axis.
 - `CFAxisVertical` can now calculate parametric coordinates for two ocean formulations (other formulations will be added as sample data becomes available to test new code on - open an [issue](https://github.com/R-CF/ncdfCF/issues) if you have such data and are looking for support). Optional use of the `units` package to deal with the various pressure units. This package is recommended if you work with data on the atmosphere or the ocean.
 - The standard names table of the CF Metadata Conventions is now accessible. The table is automatically downloaded and made available when first used; it will not be downloaded or loaded into memory when not accessed. Find standard names using the `CF$standard_names$find()` method. The table (currently 4.3MB) will be stored in the local cache of the ncdfCF package and periodically updated with the latest version.
 - New `CFAxis$coordinate_range` field to retrieve the range of the coordinates of the axis.